+++ front/src/App.js
@@ -1,11 +1,11 @@
 import "./App.css";
-import { Form } from "./components/Form";
-import { Header } from "./components/Header";
-import { Footer } from "./components/Footer";
-import { Ranking } from "./components/Ranking";
-import { About } from "./components/About";
+import { Form } from "./components/index/Form";
+import { Header } from "./components/layouts/Header";
+import { Footer } from "./components/layouts/Footer";
+import { Ranking } from "./components/ranking/Ranking";
+import { About } from "./components/about/About";
 import { BrowserRouter, Routes, Route } from "react-router-dom";
-import { Result } from "./components/Result";
+import { Result } from "./components/index/Result";
 
 function App() {
   return (
