+++ front/src/components/layouts/Footer.jsx
@@ -1,4 +1,4 @@
-import "../App.css";
+import "../../App.css";
 export const Footer = () => {
 
   return (
