+++ front/src/styles/about/About.css
